--- crates/op-rpc-types/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "op-rpc-types"
-description = "Optimism RPC types"
-
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-authors.workspace = true
-repository.workspace = true
-exclude.workspace = true
-
-[dependencies]
-# alloy-rlp = { workspace = true, features = ["arrayvec", "derive"] }
-alloy-primitives = { workspace = true, features = ["rlp", "serde", "std"] }
-# alloy-serde.workspace = true
-# alloy-genesis.workspace = true
-
-# alloy-consensus = { workspace = true, features = ["std", "serde"] }
-# alloy-eips = { workspace = true, features = ["std", "serde"] }
-
-serde = { workspace = true, features = ["derive"] }
-alloy = { workspace = true , features = ["serde", "rpc-types", "rpc-types-eth", "rlp", "consensus"]}
-serde_json.workspace = true
-
-# arbitrary
-arbitrary = { version = "1.3", features = ["derive"], optional = true }
-proptest = { version = "1.4", optional = true }
-proptest-derive = { version = "0.4", optional = true }
-
-[features]
-arbitrary = [
-    "dep:arbitrary",
-    "dep:proptest-derive",
-    "dep:proptest",
-    "alloy-primitives/arbitrary",
-]
-
-[dev-dependencies]
-arbitrary = { workspace = true, features = ["derive"] }
-proptest.workspace = true
-proptest-derive.workspace = true
-rand.workspace = true
