--- crates/op-rpc-types/src/lib.rs
@@ -1,1 +1,0 @@
-mod op;
