--- crates/op-rpc-types/src/op/mod.rs
@@ -1,5 +1,0 @@
-mod transaction;
-use crate::op::transaction::Transaction;
-use alloy::rpc::types::eth::FilterChanges as EthFilterChanges;
-
-pub type FilterChanges = EthFilterChanges<Transaction>;
