+++ crates/rpc-types/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "alloy-op-rpc-types"
+description = "Optimism RPC types"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+authors.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+[dependencies]
+alloy-primitives = { workspace = true, features = ["rlp", "serde", "std"] }
+alloy-consensus = { workspace = true, features = ["std", "serde"] }
+alloy-rpc-types.workspace = true
+
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+
+# arbitrary
+arbitrary = { version = "1.3", features = ["derive"], optional = true }
+
+[dev-dependencies]
+alloy-primitives = { workspace = true, features = ["arbitrary"] }
+alloy-consensus = { workspace = true, features = ["arbitrary"] }
+alloy-rpc-types = { workspace = true, features = ["arbitrary"] }
+
+arbitrary = { workspace = true, features = ["derive"] }
+rand.workspace = true
+
+[features]
+arbitrary = [
+    "dep:arbitrary",
+    "alloy-primitives/arbitrary",
+    "alloy-consensus/arbitrary",
+    "alloy-rpc-types/arbitrary",
+]
