+++ crates/rpc-types/README.md
@@ -1,0 +1,3 @@
+# alloy-rpc-types
+
+Ethereum RPC-related types for Alloy.
