+++ crates/rpc-types/src/transaction/mod.rs
@@ -1,0 +1,31 @@
+//! Optimism specific types related to transactions.
+
+use alloy_primitives::B256;
+use serde::{Deserialize, Serialize};
+
+pub use self::tx_type::TxType;
+
+pub mod tx_type;
+
+/// OP Transaction type
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(any(test, feature = "arbitrary"), derive(arbitrary::Arbitrary))]
+#[serde(rename_all = "camelCase")]
+pub struct Transaction {
+    /// Ethereum Transaction Types
+    #[serde(flatten)]
+    pub inner: alloy_rpc_types::Transaction,
+    /// The ETH value to mint on L2
+    #[serde(rename = "mint", skip_serializing_if = "Option::is_none")]
+    pub mint: Option<u128>,
+    /// Hash that uniquely identifies the source of the deposit.
+    #[serde(rename = "sourceHash", skip_serializing_if = "Option::is_none")]
+    pub source_hash: Option<B256>,
+    /// Field indicating whether the transaction is a system transaction, and therefore
+    /// exempt from the L2 gas limit.
+    #[serde(rename = "isSystemTx", skip_serializing_if = "Option::is_none")]
+    pub is_system_tx: Option<bool>,
+    /// Deposit receipt version for deposit transactions post-canyon
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub deposit_receipt_version: Option<u64>,
+}
